added ViewOptions
