+++ AudioCuesheetEditor/Model/Options/ApplicationOptions.cs
@@ -25,12 +25,6 @@ using System.Text.Json.Serialization;
 
 namespace AudioCuesheetEditor.Model.Options
 {
-    public enum ViewMode
-    {
-        DetailView = 0,
-        RecordView = 1,
-        ImportView = 2
-    }
     public class ApplicationOptions : Validateable, IOptions
     {
         public const LogLevel DefaultLogLevel = LogLevel.Information;
@@ -97,23 +91,7 @@ namespace AudioCuesheetEditor.Model.Options
                 }
             }
         }
-        [JsonIgnore]
-        public ViewMode ActiveTab { get; set; }
-        public String? ActiveTabName
-        {
-            get => Enum.GetName(ActiveTab);
-            set
-            {
-                if (value != null)
-                {
-                    ActiveTab = Enum.Parse<ViewMode>(value);
-                }
-                else
-                {
-                    throw new ArgumentNullException(nameof(value));
-                }
-            }
-        }
+        
         public String? ProjectFilename
         { 
             get => projectFilename;
